+++ src/JsonApi/Response/JsonApiResponse.php
@@ -3,7 +3,7 @@ namespace WoohooLabs\Yang\JsonApi\Response;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\StreamInterface;
-use WoohooLabs\Yang\JsonApi\Deserializer\DefaultDeserializer;
+use WoohooLabs\Yang\JsonApi\Deserializer\JsonDeserializer;
 use WoohooLabs\Yang\JsonApi\Deserializer\DeserializerInterface;
 use WoohooLabs\Yang\JsonApi\Schema\Document;
 
@@ -30,7 +30,7 @@ class JsonApiResponse implements ResponseInterface
     public function __construct(ResponseInterface $response, DeserializerInterface $deserializer = null)
     {
         $this->response = $response;
-        $this->deserializer = $deserializer !== null ? $deserializer : new DefaultDeserializer();
+        $this->deserializer = $deserializer !== null ? $deserializer : new JsonDeserializer();
     }
 
     /**